Windows: modify PopulateCurrentEntry follow mp rules
It is always safe to read the mount point target data, it is not
always safe to evaluate the volume id since the VLDB might not be
reachable.

For directory enumerations do not evaluate the target Fid.  Only
do so for eval by name and by id requests.

Reviewed-on: http://gerrit.openafs.org/10153
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>
(cherry picked from commit 0edbeb4a3ce86387dbc31b1745e5517accdb4823)

Change-Id: Ifa80adce31b003f015ae4d9f4d1333109f93ee0f
Reviewed-on: http://gerrit.openafs.org/10163
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>