Convert buserver to libutil's logging
The primary motivation is to get automatic log rotation, but we
also get built-in locking for the logging calls.

This does have a side effect of changing the format used to print
timestamps in the log file.

Export WriteLogBuffer from liboafs_util so as to be mostly compatible
with the ... idiosyncrasy of LogError()'s previous behavior.

Garbage-collect the unused (and un-exported) printHashTable() and LogNetDump().

Change-Id: I860370e60082ea355806b3f1945eee76db7c32e3
Reviewed-on: http://gerrit.openafs.org/10333
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Karl Ramm <kcr@1ts.org>
Reviewed-by: Derrick Brashear <shadow@your-file-system.com>