Tidy up MakefileProto.FBSD.in
We don't support FreeBSD 6.x and older now.  As such, we can convert
entirely to the "new" world order, and not have to worry about adding
new sysnames to the conditionals every time we add them.  FreeBSD 7.x
is no longer supported by upstream, but it does not impose an undue
burden on us to continue to build on FreeBSD 7.1 and higher; the code
needed for 7.0 exactly can be removed, though.

(The sense of the conditionals should really have been reversed anyway,
so that the default case would catch new sysnames correctly.  Too late now.)

This is a 1.6-only change, as the libafs build system on master is quite
different.

Change-Id: Iad5d8b6edb9140b38568ab57ffecf0cfdaf1cbd3
Reviewed-on: http://gerrit.openafs.org/10375
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>
Reviewed-by: Stephan Wiesand <stephan.wiesand@desy.de>